+++ src/Submissions/Api/Services/Directory/BiobankService.cs
@@ -1,3 +1,4 @@
+using System;
 using Biobanks.Data.Entities;
 using Biobanks.Entities.Data;
 using Biobanks.Submissions.Api.Services.Directory.Contracts;
@@ -16,19 +17,21 @@ namespace Biobanks.Submissions.Api.Services.Directory
         private readonly IGenericEFRepository<OrganisationServiceOffering> _organisationServiceOfferingRepository;
         private readonly IGenericEFRepository<OrganisationUser> _organisationUserRepository;
         private readonly UserManager<ApplicationUser> _userManager;
+        private readonly IGenericEFRepository<TokenValidationRecord> _tokenValidationRecordRepository;
+        private readonly IGenericEFRepository<TokenIssueRecord> _tokenIssueRecordRepository;
 
 
         public BiobankService(
             IGenericEFRepository<Organisation> organisationRepository,
             IGenericEFRepository<OrganisationServiceOffering> organisationServiceOfferingRepository,
             IGenericEFRepository<OrganisationUser> organisationUserRepository,
-            UserManager<ApplicationUser> userManager
-            )
+            UserManager<ApplicationUser> userManager, IGenericEFRepository<TokenIssueRecord> tokenIssueRecordRepository)
         {
             _organisationRepository = organisationRepository;
             _organisationServiceOfferingRepository = organisationServiceOfferingRepository;
             _organisationUserRepository = organisationUserRepository;
             _userManager = userManager;
+            _tokenIssueRecordRepository = tokenIssueRecordRepository;
         }
 
         public async Task<IEnumerable<OrganisationServiceOffering>> ListBiobankServiceOfferingsAsync(int biobankId)
@@ -56,5 +59,49 @@ namespace Biobanks.Submissions.Api.Services.Directory
                     x => x.Funders))
                 .Select(x => x.Funders)
                 .FirstOrDefault();
+        
+        public async Task<IEnumerable<ApplicationUser>> ListSoleBiobankAdminIdsAsync(int biobankId)
+        {
+          // Returns users who have admin role only for this biobank
+          // TODO remove the generic repo when upgrading to netcore, as it doesn't support groupby fully
+          var admins = await _organisationUserRepository.ListAsync(false);
+          var adminIds = admins.GroupBy(a => a.OrganisationUserId)
+            .Where(g => g.Count() == 1)
+            .Select(a => a.FirstOrDefault(ai => ai.OrganisationId == biobankId))
+            .Select(ou => ou?.OrganisationUserId);
+
+          return await _userManager.Users.Where(x => adminIds.Contains(x.Id)).ToListAsync();
+        }
+        
+        public async Task<string> GetUnusedTokenByUser(string biobankUserId)
+        {
+          // Check most recent token record
+          var tokenIssue = (await _tokenIssueRecordRepository.ListAsync(
+            false,
+            x => x.UserId.Contains(biobankUserId),
+            x => x.OrderBy(c => c.IssueDate))).FirstOrDefault();
+
+          // Check validation records
+          var tokenValidation = await _tokenValidationRecordRepository.ListAsync(
+            false,
+            x => x.UserId.Contains(biobankUserId));
+
+          List<string> token = tokenValidation.Select(t => t.Token).ToList();
+          DateTime now = DateTime.Now;
+
+          var user = await _userManager.FindByIdAsync(biobankUserId) ??
+                     throw new InvalidOperationException(
+                       $"Account could not be confirmed. User not found! User ID: {biobankUserId}");
+
+
+          if (tokenIssue.Equals(null) || token.Contains(tokenIssue.Token) || tokenIssue.IssueDate < now.AddHours(-20))
+          {
+            return await _userManager.GeneratePasswordResetTokenAsync(user);
+          }
+          else
+          {
+            return tokenIssue.Token;
+          }
+        }
     }
 }
